+++ app/src/main/java/com/shub39/rush/domain/util/TTMLParser.kt
@@ -1,25 +1,12 @@
-/*
- * Copyright (C) 2026  Shubham Gorai
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-package com.shub39.rush.domain.dataclasses
+package com.shub39.rush.domain.util
 
 import android.annotation.SuppressLint
-import javax.xml.parsers.DocumentBuilderFactory
+import com.shub39.rush.domain.dataclasses.ParsedLine
+import com.shub39.rush.domain.dataclasses.ParsedWord
+import com.shub39.rush.domain.dataclasses.SpanInfo
 import org.w3c.dom.Element
 import org.w3c.dom.Node
+import javax.xml.parsers.DocumentBuilderFactory
 
 object TTMLParser {
